refactor to fix typing issues
