+++ src/snowflake/snowpark/mock/plan_builder.py
@@ -2,11 +2,10 @@
 # Copyright (c) 2012-2023 Snowflake Computing Inc. All rights reserved.
 #
 
-from typing import Dict, Iterable, List, Optional, Tuple
+from typing import Dict, List, Optional, Tuple
 
 from snowflake.snowpark._internal.analyzer.expression import Attribute
 from snowflake.snowpark._internal.analyzer.snowflake_plan import SnowflakePlanBuilder
-from snowflake.snowpark._internal.type_utils import ColumnOrName
 from snowflake.snowpark._internal.utils import is_single_quoted
 from snowflake.snowpark.mock.plan import MockExecutionPlan, MockFileOperation
 
@@ -26,7 +25,8 @@ class MockSnowflakePlanBuilder(SnowflakePlanBuilder):
         schema: List[Attribute],
         schema_to_cast: Optional[List[Tuple[str, str]]] = None,
         transformations: Optional[List[str]] = None,
-        metadata_columns: Optional[Iterable[ColumnOrName]] = None,
+        metadata_project: Optional[List[str]] = None,
+        metadata_schema: Optional[List[Attribute]] = None,
     ) -> MockExecutionPlan:
         if format.upper() != "CSV":
             raise NotImplementedError(
